refactor(acr-080): retire DomainError base class and generic subclasses
Delete legacy DomainError + isDomainError entirely (313-line domain-error.ts).
BusinessRuleViolationError migrated to ResultErrorException (3 consumers
unchanged, instanceof preserved). 6 dead generic classes (NotFound/Validation/
Unauthorized/Forbidden/Conflict/InternalServerError) removed; specs use
equivalent ResultErrorException. extractErrorInfo moved & de-branched.
result-error-mapper + task-write-support isDomainError branches removed.

DOMAIN_ERROR_SUBCLASS 7 -> 0; total 28 -> 21. ACR-080 (remove DomainError)
complete. Scanner passes.
