+++ packages/task/src/server/application/use-cases/commands/task-write-support.ts
@@ -1,7 +1,7 @@
 import { ResultErrorException, type ResultError } from '@memoflow/contracts/result';
 import type { ITaskInstanceRepository } from '../../../domain/repositories/i-task-instance-repository';
 import type { ITaskTemplateRepository } from '../../../domain/repositories/i-task-template-repository';
-import { isDomainError, mapInfraErrorToResultError } from '@memoflow/utils/errors';
+import { mapInfraErrorToResultError } from '@memoflow/utils/errors';
 
 export interface TaskWriteRepositories {
   /** 完整事务（complete 等）需要模板读取；仅实例操作（uncomplete）可省略。 */
@@ -25,7 +25,7 @@ export function mapTaskWriteErrorToResultError(
   error: unknown,
   fallbackMessage: string,
 ): ResultError {
-  if (isDomainError(error) || error instanceof ResultErrorException) {
+  if (error instanceof ResultErrorException) {
     const ctx = (error as { context?: Record<string, unknown> }).context;
     const cause = (error as { cause?: unknown }).cause ?? error;
     return { code: 'BAD_REQUEST', message: error.message, context: ctx, cause };
