+++ packages/utils/src/errors/business-rule-violation-error.ts
@@ -1,0 +1,11 @@
+import { ResultErrorException } from '@memoflow/contracts/result';
+
+/**
+ * 业务规则违规错误
+ * 当领域操作违反业务规则时抛出
+ */
+export class BusinessRuleViolationError extends ResultErrorException {
+  constructor(message: string, context?: Record<string, unknown>) {
+    super(message, 'BUSINESS_ERROR', undefined, context, 400);
+  }
+}
