+++ packages/utils/src/errors/index.ts
@@ -3,7 +3,8 @@
  * 导出所有错误类型
  */
 
-export * from './domain-error';
+export * from './business-rule-violation-error';
 export * from './diagnostic-failure';
+export * from './extract-error-info';
 export * from './prisma-error-mapper';
 export * from './result-error-mapper';
