+++ packages/utils/src/result/ipc-adapter.spec.ts
@@ -13,7 +13,6 @@ import {
   ok,
 } from '@memoflow/contracts/result';
 import type { ExecutionContext, RequestContext } from '@memoflow/contracts/shared';
-import { ConflictError } from '../errors/domain-error';
 
 const AdapterFailureRegistry = defineFailureRegistry({
   TEST_PROVIDER_UNAVAILABLE: {
@@ -161,7 +160,13 @@ describe('ipcAdapter', () => {
     const controllerFn = vi
       .fn()
       .mockRejectedValue(
-        new ConflictError('Multiple repositories found', { count: 2, repositoryIds: ['repo-1'] }),
+        new ResultErrorException(
+          'Multiple repositories found',
+          'CONFLICT',
+          undefined,
+          { count: 2, repositoryIds: ['repo-1'] },
+          409,
+        ),
       );
     const handler = ipcAdapter(controllerFn, { extractContext: () => fullContext() });
